# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserShape"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserShape"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/users/me/special-needs": {
      "get": {
        "operationId": "UsersController_getSpecialNeeds",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecialNeedsShape"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "UsersController_replaceSpecialNeeds",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpecialNeedsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecialNeedsShape"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/users": {
      "get": {
        "operationId": "UsersController_listByRole",
        "parameters": [
          {
            "name": "role",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "agencyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserSummaryShape"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/users/{id}/suspend": {
      "patch": {
        "operationId": "UsersController_suspend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSummaryShape"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/users/{id}/reactivate": {
      "patch": {
        "operationId": "UsersController_reactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSummaryShape"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/rites/sheets": {
      "get": {
        "operationId": "RitesController_listPublished",
        "parameters": [
          {
            "name": "pilgrimageType",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RiteSheetShape"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "rites"
        ]
      },
      "post": {
        "operationId": "RitesController_createSheet",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRiteSheetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiteSheetShape"
                }
              }
            }
          }
        },
        "tags": [
          "rites"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/rites/sheets/admin": {
      "get": {
        "operationId": "RitesController_listAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RiteSheetShape"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "rites"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/rites/sheets/{id}": {
      "patch": {
        "operationId": "RitesController_updateSheet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRiteSheetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiteSheetShape"
                }
              }
            }
          }
        },
        "tags": [
          "rites"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/rites/sheets/{id}/validate": {
      "patch": {
        "operationId": "RitesController_validateSheet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiteSheetShape"
                }
              }
            }
          }
        },
        "tags": [
          "rites"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/rites/progress": {
      "get": {
        "operationId": "RitesController_getMyProgress",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RiteProgressShape"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "rites"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/rites/progress/sync": {
      "post": {
        "operationId": "RitesController_syncProgress",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRiteProgressDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RiteProgressShape"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "rites"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/rites/progress/{riteKey}/reset-counter": {
      "patch": {
        "operationId": "RitesController_resetCounter",
        "parameters": [
          {
            "name": "riteKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiteProgressShape"
                }
              }
            }
          }
        },
        "tags": [
          "rites"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Plateforme Oumra & Hadj — API",
    "description": "API centrale : pèlerins, agences, forfaits, réservations, paiements, documents, rites, groupes, notifications.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "EmergencyContactShape": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "relationship": {
            "type": "string"
          }
        },
        "required": [
          "fullName",
          "phone"
        ]
      },
      "UserShape": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "enum": [
              "pilgrim",
              "agency",
              "guide",
              "admin"
            ],
            "type": "string"
          },
          "preferredLanguage": {
            "type": "string"
          },
          "emergencyContact": {
            "$ref": "#/components/schemas/EmergencyContactShape"
          },
          "bloodType": {
            "type": "string"
          },
          "passportNumber": {
            "type": "string"
          },
          "agencyId": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "fullName",
          "role",
          "preferredLanguage",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "EmergencyContactDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "relationship": {
            "type": "string"
          }
        }
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "preferredLanguage": {
            "type": "string",
            "enum": [
              "fr",
              "en",
              "ar"
            ]
          },
          "emergencyContact": {
            "$ref": "#/components/schemas/EmergencyContactDto"
          },
          "bloodType": {
            "type": "string"
          },
          "passportNumber": {
            "type": "string"
          }
        }
      },
      "SpecialNeedsShape": {
        "type": "object",
        "properties": {
          "mobility": {
            "type": "object"
          },
          "dietary": {
            "type": "string"
          },
          "medical": {
            "type": "string"
          },
          "assistance": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "mobility"
        ]
      },
      "UpdateSpecialNeedsDto": {
        "type": "object",
        "properties": {
          "mobility": {
            "type": "string",
            "enum": [
              "none",
              "reduced",
              "wheelchair"
            ]
          },
          "dietary": {
            "type": "string",
            "maxLength": 500,
            "description": "Régime alimentaire"
          },
          "medical": {
            "type": "string",
            "maxLength": 500,
            "description": "Traitement, allergie, condition à connaître"
          },
          "assistance": {
            "type": "string",
            "maxLength": 500,
            "description": "Accompagnement souhaité"
          }
        },
        "required": [
          "mobility"
        ]
      },
      "UserSummaryShape": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "enum": [
              "pilgrim",
              "agency",
              "guide",
              "admin"
            ],
            "type": "string"
          },
          "agencyId": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "fullName",
          "role",
          "isActive",
          "createdAt"
        ]
      },
...
      "RiteSheetShape": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "pilgrimageType": {
            "enum": [
              "oumra",
              "hadj",
              "both"
            ],
            "type": "string"
          },
          "order": {
            "type": "number"
          },
          "content": {
            "type": "string"
          },
          "audioRef": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "version": {
            "type": "number"
          },
          "isValidated": {
            "type": "boolean"
          },
          "validatedById": {
            "type": "string"
          },
          "validatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "key",
          "title",
          "pilgrimageType",
          "order",
          "content",
          "language",
          "version",
          "isValidated"
        ]
      },
      "CreateRiteSheetDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "minLength": 2
          },
          "title": {
            "type": "string",
            "minLength": 2
          },
          "pilgrimageType": {
            "enum": [
              "oumra",
              "hadj",
              "both"
            ],
            "type": "string"
          },
          "order": {
            "type": "number",
            "minimum": 0
          },
          "content": {
            "type": "string",
            "minLength": 10
          },
          "audioRef": {
            "type": "string"
          },
          "language": {
            "type": "string",
            "enum": [
              "fr",
              "en",
              "ar"
            ]
          }
        },
        "required": [
          "key",
          "title",
          "pilgrimageType",
          "order",
          "content"
        ]
      },
      "UpdateRiteSheetDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "minLength": 2
          },
          "title": {
            "type": "string",
            "minLength": 2
          },
          "pilgrimageType": {
            "enum": [
              "oumra",
              "hadj",
              "both"
            ],
            "type": "string"
          },
          "order": {
            "type": "number",
            "minimum": 0
          },
          "content": {
            "type": "string",
            "minLength": 10
          },
          "audioRef": {
            "type": "string"
          },
          "language": {
            "type": "string",
            "enum": [
              "fr",
              "en",
              "ar"
            ]
          }
        }
      },
      "RiteProgressShape": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "pilgrimId": {
            "type": "string"
          },
          "riteKey": {
            "type": "string"
          },
          "completed": {
            "type": "boolean"
          },
          "tawafCount": {
            "type": "number"
          },
          "saiCount": {
            "type": "number"
          },
          "clientUpdatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "pilgrimId",
          "riteKey",
          "completed",
          "tawafCount",
          "saiCount",
          "clientUpdatedAt"
        ]
      },
      "RiteProgressItemDto": {
        "type": "object",
        "properties": {
          "riteKey": {
            "type": "string",
            "minLength": 2
          },
          "completed": {
            "type": "boolean"
          },
          "tawafCount": {
            "type": "number",
            "minimum": 0
          },
          "saiCount": {
            "type": "number",
            "minimum": 0
          },
          "clientUpdatedAt": {
            "type": "string"
          }
        },
        "required": [
          "riteKey",
          "clientUpdatedAt"
        ]
      },
      "SyncRiteProgressDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RiteProgressItemDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
...
    }
  }
}