# src/modules/users/users.controller.ts
import {
  Body,
  Controller,
  Get,
  Param,
  Patch,
  Put,
  Query,
} from '@nestjs/common';
import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
import { CurrentUser } from '../../common/decorators/current-user.decorator';
import { Roles } from '../../common/decorators/roles.decorator';
import { Role } from '../../common/enums/role.enum';
import { JwtPayload } from '../../common/interfaces/authenticated-request.interface';
import {
  SpecialNeedsShape,
  UserShape,
  UserSummaryShape,
} from '../../types/user.types';
import { UpdateProfileDto } from './dto/update-profile.dto';
import { UpdateSpecialNeedsDto } from './dto/update-special-needs.dto';
import { UsersService } from './users.service';
import { Audited } from '../audit/audited.decorator';

@ApiBearerAuth()
@ApiTags('users')
@Controller('users')
export class UsersController {
  constructor(private readonly usersService: UsersService) {}

  @Get('me')
  getProfile(@CurrentUser() user: JwtPayload): Promise<UserShape> {
    return this.usersService.findByIdOrFail(user.sub);
  }

  @Patch('me')
  updateProfile(
    @CurrentUser() user: JwtPayload,
    @Body() dto: UpdateProfileDto,
  ): Promise<UserShape> {
    return this.usersService.updateProfile(user.sub, dto);
  }

  // Idée #69 — besoins spéciaux du pèlerin, déclarés par lui seul.
  @Roles(Role.PILGRIM)
  @Get('me/special-needs')
  getSpecialNeeds(@CurrentUser() user: JwtPayload): Promise<SpecialNeedsShape> {
    return this.usersService.getSpecialNeeds(user.sub);
  }

  @Roles(Role.PILGRIM)
  @Put('me/special-needs')
  replaceSpecialNeeds(
    @CurrentUser() user: JwtPayload,
    @Body() dto: UpdateSpecialNeedsDto,
  ): Promise<SpecialNeedsShape> {
    return this.usersService.replaceSpecialNeeds(user.sub, dto);
  }

  // Réservé à l'admin plateforme (supervision des utilisateurs — cahier des
  // charges §3.4 "Gestion des utilisateurs").
  @Roles(Role.ADMIN)
  @Get()
  listByRole(
    @Query('role') role: Role,
    @Query('agencyId') agencyId?: string,
  ): Promise<UserSummaryShape[]> {
    return this.usersService.listForAdmin(role, agencyId);
  }

  @Roles(Role.ADMIN)
  @Audited({ action: 'user.suspend', entityType: 'user' })
  @Patch(':id/suspend')
  suspend(@Param('id') id: string): Promise<UserSummaryShape> {
    return this.usersService.setActiveForAdmin(id, false);
  }

  @Roles(Role.ADMIN)
  @Audited({ action: 'user.reactivate', entityType: 'user' })
  @Patch(':id/reactivate')
  reactivate(@Param('id') id: string): Promise<UserSummaryShape> {
    return this.usersService.setActiveForAdmin(id, true);
  }
}